Adding solutions to CSharp technical interview questions
OddSum:  Given an array of integers, write a method to total the odd numbers.
ArraySum:  Write a method to sum the elements in an array (elements may be very large).
StringReverse:  Given a string, reverse it.
DeleteRepeat:  Given a string, remove any repeated letters.
FizzBuzz:  For a range of integers, print "Fizz" if divisible by 3, print "Buzz" if divisible by 5, and print "FizzBuzz" if the number is divisible by both 3 and 5.